Renaming subcomponents and folders
